Updating readme with right numbers

+++ compile_images.py
@@ -52,7 +52,7 @@ for i, object_url_string in tqdm(
             break
 
 success_coadd_ids = coadd_ids[~failed_mask]
-success_imgs_array = imgs[~failed_mask].astype("float32") # Compressing from float64 for storage purposes
+success_imgs_array = imgs[~failed_mask]
 np.savez_compressed(
     "./data/processed/ids_images.npz", ids=success_coadd_ids, imgs=success_imgs_array
 )
